# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for SUT SE 2563",
        "title": "SUT SA Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/districts": {
            "get": {
                "description": "list district entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List district entities",
                "operationId": "list-district",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.District"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create district",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create district",
                "operationId": "create-district",
                "parameters": [
                    {
                        "description": "district entity",
                        "name": "resolution",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.District"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.District"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/districts/{id}": {
            "get": {
                "description": "get district by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a district entity by ID",
                "operationId": "get-district",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "District ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.District"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
        "/subdistricts": {
            "get": {
                "description": "list subdistrict entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List subdistrict entities",
                "operationId": "list-subdistrict",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Subdistrict"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create subdistrict",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create subdistrict",
                "operationId": "create-subdistrict",
                "parameters": [
                    {
                        "description": "subdistrict entity",
                        "name": "resolution",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Subdistrict"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Subdistrict"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/subdistricts/{id}": {
            "get": {
                "description": "get subdistrict by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a subdistrict entity by ID",
                "operationId": "get-subdistrict",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Subdistrict ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Subdistrict"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "ent.District": {
            "type": "object",
            "properties": {
                "district": {
                    "description": "District holds the value of the \"district\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the DistrictQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DistrictEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.DistrictEdges": {
            "type": "object",
            "properties": {
                "province": {
                    "description": "Province holds the value of the province edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Province"
                    }
                }
            }
        },
        "ent.Province": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ProvinceQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.ProvinceEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                }
            }
        },
        "ent.ProvinceEdges": {
            "type": "object",
            "properties": {
                "district": {
                    "description": "District holds the value of the district edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.District"
                },
                "student": {
                    "description": "Student holds the value of the student edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Student"
                },
                "subdistrict": {
                    "description": "Subdistrict holds the value of the subdistrict edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Subdistrict"
                }
            }
        },
        "ent.Student": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the StudentQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.StudentEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "student": {
                    "description": "Student holds the value of the \"student\" field.",
                    "type": "string"
                }
            }
        },
        "ent.StudentEdges": {
            "type": "object",
            "properties": {
                "province": {
                    "description": "Province holds the value of the province edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Province"
                    }
                }
            }
        },
        "ent.Subdistrict": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the SubdistrictQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.SubdistrictEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "subdistrict": {
                    "description": "Subdistrict holds the value of the \"subdistrict\" field.",
                    "type": "string"
                }
            }
        },
        "ent.SubdistrictEdges": {
            "type": "object",
            "properties": {
                "province": {
                    "description": "Province holds the value of the province edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Province"
                    }
                }
            }
        },
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
...
}